+++ change-log.md
@@ -3,3 +3,4 @@
 ## Fixes
 * Recommended weapons not consistently rendering tabs as rarity hierarchy
 * Nod-Krai filter not being in SearchableMobList
+* Sigewinne's bonus talents not showing off-field damage